Extract `time_function` to utility.timing
